refactor(node-ui): extract PCA client + shared HTTP transport, split 26-pca.css (#1345, #1354)
Pure relocation, no behavior change — two deferred structural cleanups from the
PCA node-UI P0 (#1343).

#1345: lift the generic fetch layer (BASE, authHeaders, HttpError, fetchWithTimeout,
get/getJson/getWithTimeout, post, put, del/delJson, the Window.__DKG_TOKEN__ aug) into
a new src/ui/http.ts shared by both api.ts and the new src/ui/api/pca.ts (the PCA
client: Pca* types, fetch/create/pca*/list* fns, error predicates, PCA_ERROR_CODES,
describePcaError). api.ts becomes a compatibility barrel (`export * from './api/pca.js'`
+ `export { authHeaders, HttpError } from './http.js'`) so all 21 consumers, the tests,
the devnet e2e spec, and api-wrapper.ts keep resolving `../api.js` unchanged. The
transport had to become shared, not PCA-private: post alone has ~24 non-PCA callers.

#1354: split the 1,135-line 26-pca.css into six surface-aligned stylesheets
(primitives / overview-cards / shell-dashboard / modals-detail / publish-eligibility /
followup-overrides), @import'd from styles.css in a cascade-preserving order (the
override tail imported last). Class names + rendered output identical.

Preserved invariants: the \x00 dedup byte stays in api.ts only; describePcaError keeps
isRpcTransportError before the capability-503 branch; ConvictionCostCovered /
PcaCostCoveredNotif deliberately stay in api.ts (publish/notifications domain); module
graph is acyclic (api.ts -> api/pca.ts -> http.ts).

Verified: test/pca-api.test.ts 28/28 through the unchanged barrel; 846 pass / 0 failed
across the full 51-file api.js blast radius; tsc -p src/ui/tsconfig.json zero error-delta
vs origin/main baseline (all three files confirmed in the program via --listFiles); vite
build:ui exit 0; the six CSS files' source-order concat is byte-identical to the original
(the sole import-order reorder is between disjoint selector sets). Adversarial gate: 4/5
lenses CLEAN, no must-fix findings.
