+++ packages/node-ui/src/ui/styles.css
@@ -25,4 +25,9 @@
 @import './styles/23-agent-activity-profile.css';
 @import './styles/24-memory-stack-identity-primer.css';
 @import './styles/25-graph-provenance-card.css';
-@import './styles/26-pca.css';
+@import './styles/26-pca-primitives.css';
+@import './styles/26-pca-overview-cards.css';
+@import './styles/26-pca-shell-dashboard.css';
+@import './styles/26-pca-modals-detail.css';
+@import './styles/26-pca-publish-eligibility.css';
+@import './styles/26-pca-followup-overrides.css';
