+++ packages/node-ui/src/ui/styles/26-pca-followup-overrides.css
@@ -1,0 +1,241 @@
+/* PCA styles — PCA UX follow-up override layer. Split from 26-pca.css (#1354); class names + cascade order unchanged. MUST import LAST — equal-specificity overrides that win by source order across overview/primitives/modals. */
+
+/* ----- PCA UX follow-up: combined coverage + compact controls ----- */
+.v10-pca-linked-pcas {
+  margin-top: 16px;
+  padding-top: 14px;
+  border-top: 1px solid var(--border-subtle);
+}
+.v10-pca-card-grid-compact {
+  grid-template-columns: repeat(auto-fill, minmax(320px, 1fr));
+  margin-top: 10px;
+}
+.v10-pca-card-summary {
+  margin: 10px 0 0;
+  font-size: 12px;
+  line-height: 1.45;
+  color: var(--text-secondary);
+}
+.v10-pca-overview-owner-actions {
+  display: flex;
+  align-items: flex-start;
+  justify-content: flex-end;
+  gap: 12px;
+  flex-wrap: wrap;
+  max-width: 100%;
+}
+.v10-pca-overview-wallet {
+  width: auto;
+  max-width: min(420px, 100%);
+  margin-left: 0;
+  margin-top: 8px;
+}
+.v10-pca-overview-wallet + .v10-pca-owned-groups {
+  margin-top: 18px;
+}
+.v10-pca-wallet-connect.compact {
+  min-width: 0;
+  max-width: 420px;
+}
+.v10-pca-wallet-connect.compact .v10-pca-wallet-note {
+  max-width: 420px;
+}
+.v10-pca-wallet-pill {
+  max-width: 100%;
+}
+.v10-pca-wallet-pill-main,
+.v10-pca-wallet-pill-provider {
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+.v10-pca-wallet-pill-wrap {
+  max-width: 100%;
+}
+.v10-pca-modal-box .v10-modal-body {
+  display: flex;
+  flex-direction: column;
+  gap: 16px;
+}
+.v10-pca-get-added-body {
+  gap: 12px;
+}
+.v10-pca-modal-box .v10-pca-detail-section {
+  margin-top: 0;
+}
+.v10-pca-get-added-body .v10-pca-detail-section:first-child {
+  border-top: 0;
+  padding-top: 0;
+}
+.v10-pca-get-added-body .v10-pca-detail-section {
+  padding-top: 14px;
+}
+.v10-pca-get-added-body .v10-pca-detail-agentlist {
+  margin-top: 8px;
+}
+.v10-pca-get-added-body .v10-pca-detail-form {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr) auto;
+  align-items: center;
+  gap: 10px;
+  max-width: 420px;
+}
+.v10-pca-get-added-body .v10-pca-detail-form .v10-form-input {
+  max-width: none;
+}
+.v10-pca-get-added-body .v10-pca-detail-form .v10-pca-card-btn.primary {
+  min-height: 36px;
+  padding-top: 5px;
+  padding-bottom: 5px;
+}
+.v10-pca-discovered-check {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+  flex-wrap: wrap;
+  margin: 8px 0;
+  font-size: 11px;
+  line-height: 1.5;
+  color: var(--text-tertiary);
+}
+.v10-pca-create-success {
+  margin: 0;
+  font-size: 12px;
+  line-height: 1.55;
+  color: var(--text-secondary);
+}
+.v10-pca-handshake-copyall {
+  margin-top: 8px;
+}
+.v10-pca-create-section {
+  margin-bottom: 0;
+}
+.v10-pca-create-section + .v10-pca-create-section {
+  padding-top: 18px;
+}
+.v10-pca-owner-key-card {
+  gap: 10px;
+  padding: 12px;
+}
+.v10-pca-owner-key-readiness .v10-pca-wallet-connect,
+.v10-pca-detail-band + .v10-modal-tip .v10-pca-wallet-pill-wrap {
+  margin-top: 8px;
+}
+.v10-pca-node-picker-combo {
+  margin-top: 10px;
+}
+.v10-pca-node-picker-combo.has-selection .v10-form-input::placeholder {
+  color: var(--text-primary);
+  font-weight: 700;
+  opacity: 1;
+}
+.v10-pca-node-picker-list {
+  list-style: none;
+  margin: 8px 0 0;
+  padding: 0;
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  max-height: 280px;
+  overflow: auto;
+}
+.v10-pca-node-picker-option {
+  display: grid;
+  grid-template-columns: minmax(64px, auto) minmax(0, 1fr) auto minmax(120px, auto) auto;
+  align-items: center;
+  gap: 8px;
+  padding: 8px 10px;
+  border: 1px solid var(--border-subtle);
+  border-radius: 7px;
+  background: var(--bg-surface);
+  cursor: pointer;
+}
+.v10-pca-node-picker-option.active,
+.v10-pca-node-picker-option.selected {
+  border-color: var(--text-link);
+  background: var(--bg-hover);
+}
+.v10-pca-node-picker-peer {
+  min-width: 0;
+  overflow: hidden;
+  text-overflow: ellipsis;
+  white-space: nowrap;
+}
+.v10-pca-node-picker-stake {
+  color: var(--text-secondary);
+  white-space: nowrap;
+}
+.v10-pca-node-picker-paste {
+  display: grid;
+  grid-template-columns: minmax(0, 1fr) auto;
+  align-items: start;
+  gap: 8px;
+  margin-top: 10px;
+}
+.v10-pca-node-picker-selected {
+  display: flex;
+  align-items: center;
+  gap: 10px;
+  flex-wrap: wrap;
+  margin-top: 8px;
+  padding: 10px 12px;
+  border: 1px solid rgba(96, 165, 250, 0.35);
+  border-radius: 8px;
+  background: rgba(96, 165, 250, 0.08);
+  color: var(--text-secondary);
+  font-size: 12px;
+  line-height: 1.4;
+}
+.v10-pca-node-picker-selected strong {
+  color: var(--text-primary);
+  font-size: 13px;
+  font-weight: 800;
+}
+.v10-pca-create-review {
+  gap: 8px;
+  max-width: 560px;
+}
+.v10-pca-create-review > div {
+  display: grid;
+  grid-template-columns: minmax(120px, 1fr) minmax(120px, auto);
+  align-items: center;
+  gap: 16px;
+}
+.v10-pca-create-owner-note {
+  margin-top: 12px;
+}
+.v10-pca-create-owner-note + .v10-modal-tip {
+  margin-top: 10px;
+}
+.v10-pca-create-section .v10-pca-node-picker {
+  margin-top: 10px;
+}
+.v10-pca-detail > .v10-pca-device-progress {
+  margin: 16px 0;
+}
+.v10-pca-detail .v10-modal-error {
+  max-width: 100%;
+  margin: 12px 0;
+  overflow-wrap: anywhere;
+}
+.v10-pca-detail .v10-pca-device-progress + .v10-stat-strip {
+  margin-top: 16px;
+}
+@media (max-width: 720px) {
+  .v10-pca-section-head > .v10-pca-overview-cta-group {
+    justify-content: flex-end;
+    margin-left: auto;
+  }
+  .v10-pca-get-added-body .v10-pca-detail-form {
+    grid-template-columns: 1fr;
+  }
+  .v10-pca-node-picker-option {
+    grid-template-columns: auto minmax(0, 1fr);
+  }
+  .v10-pca-node-picker-option .badge,
+  .v10-pca-node-picker-stake,
+  .v10-pca-node-picker-own {
+    grid-column: 2;
+  }
+}
