+++ packages/node-ui/src/ui/styles/26-pca-modals-detail.css
@@ -1,0 +1,129 @@
+/* PCA styles — S3 Detail + S2 Create modal + S4 Approve modal. Split from 26-pca.css (#1354); class names + cascade order unchanged. Groups the detail scaffold (was above the disjoint dashboard block) with the rest of detail/modals. */
+
+/* ----- S3 Detail (scaffold) ----- */
+.v10-pca-detail-band { display: flex; align-items: center; gap: 12px; margin-bottom: 16px; }
+.v10-pca-detail-id { font-size: 16px; font-weight: 700; color: var(--text-primary); }
+.v10-pca-detail-committed { font-size: 12px; color: var(--text-secondary); }
+.v10-pca-detail-owner { display: flex; align-items: center; gap: 8px; margin-top: 16px; }
+.v10-pca-detail-note {
+  margin-top: 16px;
+  font-size: 12px;
+  color: var(--text-tertiary);
+  line-height: 1.5;
+}
+.v10-pca-detail .v10-stat-strip-value {
+  min-height: 42px;
+}
+
+/* ----- S2 Create modal (C1) ----- */
+.v10-pca-modal-box { max-width: 560px; }
+.v10-pca-create-section { margin-bottom: 18px; }
+.v10-pca-create-section + .v10-pca-create-section { border-top: 1px solid var(--border-subtle); padding-top: 16px; }
+.v10-pca-create-section-title {
+  font-size: 12px;
+  font-weight: 700;
+  color: var(--text-primary);
+  margin: 0 0 10px;
+}
+.v10-pca-create-balance { font-size: 11px; color: var(--text-secondary); margin: 6px 0 0; }
+.v10-pca-create-err { color: var(--text-danger); }
+.v10-pca-create-hint { font-size: 11px; color: var(--text-tertiary); margin: 6px 0 0; line-height: 1.45; }
+.v10-pca-create-warn {
+  margin: 10px 0 0;
+  padding: 8px 10px;
+  font-size: 11px;
+  line-height: 1.5;
+  color: var(--text-warning);
+  background: var(--amber-dim);
+  border-radius: 6px;
+}
+.v10-pca-create-tx { font-size: 11px; color: var(--text-secondary); margin: 8px 0 0; word-break: break-all; }
+.v10-pca-create-review { display: flex; flex-direction: column; gap: 6px; }
+.v10-pca-create-review > div { display: flex; justify-content: space-between; font-size: 12px; }
+.v10-pca-create-review span { color: var(--text-tertiary); }
+.v10-pca-create-review strong { color: var(--text-primary); }
+.v10-pca-create-owner { display: flex; align-items: center; gap: 8px; margin-top: 12px; }
+.v10-pca-create-owner-note { font-size: 11px; color: var(--text-tertiary); margin: 6px 0 0; line-height: 1.45; }
+.v10-pca-create-success-stats { display: flex; flex-direction: column; gap: 4px; font-size: 13px; color: var(--text-secondary); margin-bottom: 10px; }
+.v10-pca-create-recon { font-size: 12px; color: var(--text-secondary); line-height: 1.5; }
+.v10-pca-create-recon code { word-break: break-all; }
+.v10-pca-create-recon-actions { margin: 8px 0 0; padding-left: 18px; }
+.v10-pca-create-recon-actions li { margin-bottom: 4px; }
+.v10-pca-modal-box .v10-modal-error,
+.v10-pca-modal-box .v10-modal-warning,
+.v10-pca-modal-box .v10-modal-tip {
+  margin: 0;
+  overflow-wrap: anywhere;
+}
+.v10-pca-owner-key-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(220px, 1fr));
+  gap: 10px;
+}
+.v10-pca-owner-key-card {
+  display: grid;
+  grid-template-columns: auto 1fr;
+  align-items: flex-start;
+  gap: 8px;
+  padding: 10px;
+  border: 1px solid var(--border-subtle);
+  border-radius: 7px;
+  background: var(--bg-surface);
+  min-width: 0;
+}
+.v10-pca-owner-key-card.selected {
+  border-color: var(--text-link);
+}
+.v10-pca-owner-key-card strong,
+.v10-pca-owner-key-card small {
+  display: block;
+  min-width: 0;
+}
+.v10-pca-owner-key-card small {
+  margin-top: 4px;
+  color: var(--text-secondary);
+  font-size: 11px;
+  line-height: 1.45;
+}
+.v10-pca-owner-key-readiness {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  margin-top: 10px;
+}
+
+/* ----- S4 Approve modal (C2) ----- */
+.v10-pca-approve-checklist { display: flex; flex-direction: column; gap: 6px; margin: 10px 0; }
+.v10-pca-approve-self-row { display: flex; align-items: center; gap: 8px; }
+.v10-pca-approve-self-row .v10-pca-wallet-row { flex: 1; }
+.v10-pca-approve-counts { font-size: 11px; color: var(--text-secondary); margin: 8px 0; display: flex; align-items: center; gap: 6px; flex-wrap: wrap; }
+.v10-pca-approve-results { display: flex; flex-direction: column; gap: 6px; margin-top: 14px; }
+.v10-pca-approve-summary { font-size: 12px; font-weight: 600; color: var(--text-primary); margin: 4px 0 0; }
+.v10-pca-approve-handoff { margin-top: 14px; display: flex; flex-direction: column; gap: 10px; }
+
+/* ----- S3 Detail (D1) ----- */
+.v10-pca-detail-section { margin-top: 18px; padding-top: 16px; border-top: 1px solid var(--border-subtle); }
+.v10-pca-detail-section-title { font-size: 13px; font-weight: 700; color: var(--text-primary); margin: 0 0 8px; }
+.v10-pca-detail-hint { font-size: 11px; color: var(--text-tertiary); line-height: 1.5; margin: 6px 0; }
+.v10-pca-detail-subhead { font-size: 11px; font-weight: 600; color: var(--text-secondary); margin: 12px 0 6px; }
+.v10-pca-detail-form { display: flex; align-items: center; gap: 8px; flex-wrap: wrap; margin: 8px 0; }
+.v10-pca-detail-form .v10-form-input,
+.v10-pca-detail-form .v10-form-select { max-width: 280px; }
+.v10-pca-detail-form .v10-pca-address-crux { flex: 1; min-width: 240px; }
+.v10-pca-detail-result { font-size: 11px; color: var(--text-success); margin: 8px 0 0; line-height: 1.5; }
+.v10-pca-detail-agentlist { display: flex; flex-direction: column; gap: 8px; }
+.v10-pca-detail-agentlist + .v10-pca-card-btn,
+.v10-pca-detail-subhead + .v10-pca-detail-agentlist + .v10-pca-card-btn {
+  margin-top: 10px;
+}
+.v10-pca-agent-row { display: flex; }
+.v10-pca-agent-row .v10-pca-wallet-row { flex: 1; }
+.v10-pca-agent-confirm {
+  display: inline-flex;
+  align-items: center;
+  gap: 6px;
+  flex-wrap: wrap;
+  font-size: 11px;
+  color: var(--text-warning);
+}
+
