+++ packages/node-ui/src/ui/styles/26-pca-overview-cards.css
@@ -1,0 +1,220 @@
+/* PCA styles — S1 Conviction Overview shell + PcaAccountCard + Track-by-id disclosure. Split from 26-pca.css (#1354); class names + cascade order unchanged. */
+
+/* ----- S1 Conviction Overview ----- */
+.v10-pca-overview { padding: 20px 24px; max-width: 980px; }
+.v10-pca-unavailable,
+.v10-pca-detail { padding: 24px; max-width: 880px; }
+.v10-pca-overview-head { margin-bottom: 16px; }
+.v10-pca-overview-titlerow { display: flex; align-items: center; gap: 12px; }
+.v10-pca-overview-title { font-size: 18px; font-weight: 700; margin: 0; color: var(--text-primary); }
+.v10-pca-role-chip { text-transform: capitalize; }
+.v10-pca-overview-explainer {
+  margin: 8px 0 0;
+  font-size: 12px;
+  line-height: 1.6;
+  color: var(--text-secondary);
+  max-width: 760px;
+}
+.v10-pca-edge-banner { margin-bottom: 16px; }
+.v10-pca-overview-toolbar {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  gap: 12px;
+  margin-bottom: 16px;
+  flex-wrap: wrap;
+}
+.v10-pca-overview-wallet { margin-left: auto; }
+.v10-pca-filter { display: inline-flex; border: 1px solid var(--border-default); border-radius: 8px; overflow: hidden; }
+.v10-pca-filter-tab {
+  border: none;
+  background: var(--bg-surface);
+  color: var(--text-secondary);
+  font-size: 12px;
+  font-weight: 600;
+  padding: 6px 14px;
+  cursor: pointer;
+}
+.v10-pca-filter-tab + .v10-pca-filter-tab { border-left: 1px solid var(--border-default); }
+.v10-pca-filter-tab.active { background: var(--bg-active); color: var(--text-primary); }
+.v10-pca-filter-tab:focus-visible { outline: 2px solid var(--text-link); outline-offset: -2px; }
+.v10-pca-overview-caveat,
+.v10-pca-overview-foot { font-size: 11px; color: var(--text-tertiary); line-height: 1.5; margin: 12px 0 0; }
+.v10-pca-section {
+  margin-top: 18px;
+}
+.v10-pca-section + .v10-pca-section {
+  margin-top: 24px;
+}
+.v10-pca-section-head {
+  display: flex;
+  align-items: flex-start;
+  justify-content: space-between;
+  gap: 16px;
+  flex-wrap: wrap;
+  margin-bottom: 12px;
+}
+.v10-pca-section-head > div:first-child {
+  flex: 1 1 320px;
+  min-width: 0;
+}
+.v10-pca-section-head > .v10-pca-overview-cta-group {
+  flex: 0 0 auto;
+  justify-content: flex-end;
+  margin-left: auto;
+}
+.v10-pca-section-title {
+  margin: 0;
+  font-size: 14px;
+  font-weight: 700;
+  color: var(--text-primary);
+}
+.v10-pca-coverage-summary {
+  padding: 14px 0 4px;
+  border-top: 1px solid var(--border-subtle);
+}
+.v10-pca-overview-cta-group {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  flex-wrap: wrap;
+}
+.v10-pca-overview-wallet {
+  width: min(460px, 100%);
+}
+.v10-pca-wallet-list {
+  display: flex;
+  flex-direction: column;
+  gap: 6px;
+  max-width: 680px;
+  margin-top: 10px;
+}
+.v10-pca-coverage-summary .v10-pca-wallet-list {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(min(100%, 320px), 380px));
+  gap: 8px;
+  max-width: none;
+  justify-content: start;
+}
+.v10-pca-inline-status {
+  margin: 12px 0 0;
+  font-size: 12px;
+  line-height: 1.5;
+  color: var(--text-secondary);
+}
+
+/* ----- PcaAccountCard ----- */
+.v10-pca-card-grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fill, minmax(440px, 1fr));
+  gap: 16px;
+}
+@media (max-width: 560px) {
+  .v10-pca-card-grid { grid-template-columns: 1fr; }
+}
+.v10-pca-card-title { display: flex; align-items: center; gap: 8px; margin-bottom: 12px; }
+.v10-pca-card-id { font-size: 14px; font-weight: 700; color: var(--text-primary); }
+.v10-pca-card-sponsor { font-size: 11px; color: var(--text-tertiary); }
+.v10-pca-card-spacer { flex: 1; }
+.v10-pca-card-owner { display: flex; align-items: center; gap: 8px; margin-top: 12px; }
+.v10-pca-card-owner-lbl { font-size: 11px; color: var(--text-tertiary); min-width: 44px; }
+.v10-pca-card-owner-note { font-size: 11px; color: var(--text-tertiary); margin: 4px 0 0; line-height: 1.45; }
+.v10-pca-card-explorer { color: var(--text-link); text-decoration: none; font-size: 13px; }
+.v10-pca-card-warn {
+  margin: 10px 0 0;
+  padding: 8px 10px;
+  font-size: 11px;
+  line-height: 1.5;
+  color: var(--text-warning);
+  background: var(--amber-dim);
+  border-radius: 6px;
+}
+.v10-pca-card-meta { font-size: 11px; color: var(--text-secondary); margin: 10px 0 0; }
+.v10-pca-card-meta-sep { color: var(--text-ghost); }
+.v10-pca-card-wallets { display: flex; flex-direction: column; gap: 6px; margin-top: 12px; }
+.v10-pca-card-actions { display: flex; flex-wrap: wrap; align-items: center; gap: 8px; margin-top: 14px; }
+.v10-pca-owned-groups,
+.v10-pca-account-group {
+  display: flex;
+  flex-direction: column;
+  gap: 14px;
+}
+.v10-pca-account-group + .v10-pca-account-group { margin-top: 18px; }
+.v10-pca-account-group-head {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+}
+.v10-pca-account-group-title {
+  margin: 0;
+  font-size: 13px;
+  font-weight: 700;
+  color: var(--text-primary);
+}
+.v10-pca-account-group-count {
+  min-width: 22px;
+  height: 20px;
+  display: inline-flex;
+  align-items: center;
+  justify-content: center;
+  border: 1px solid var(--border-default);
+  border-radius: 999px;
+  font-size: 11px;
+  color: var(--text-secondary);
+}
+.v10-pca-card-btn {
+  border: 1px solid var(--border-default);
+  background: var(--bg-surface);
+  color: var(--text-secondary);
+  font-size: 12px;
+  font-weight: 600;
+  padding: 6px 12px;
+  border-radius: 7px;
+  cursor: pointer;
+}
+.v10-pca-card-btn:hover:not(:disabled) { background: var(--bg-hover); color: var(--text-primary); }
+.v10-pca-card-btn.compact {
+  padding: 4px 10px;
+  min-height: 32px;
+}
+.v10-pca-card-btn.primary {
+  background: var(--accent-blue);
+  border-color: var(--accent-blue);
+  color: var(--text-on-accent);
+}
+.v10-pca-card-btn:disabled { opacity: 0.5; cursor: not-allowed; }
+.v10-pca-card-btn:focus-visible { outline: 2px solid var(--text-link); outline-offset: 1px; }
+.v10-pca-inline-btn {
+  border: none;
+  background: transparent;
+  color: var(--text-link);
+  font: inherit;
+  font-weight: 700;
+  padding: 0;
+  cursor: pointer;
+  text-decoration: underline;
+}
+.v10-pca-inline-btn:focus-visible { outline: 2px solid var(--text-link); outline-offset: 1px; }
+.v10-pca-card-error { display: flex; flex-direction: column; gap: 10px; font-size: 12px; color: var(--text-secondary); }
+.v10-pca-card-skeleton { min-height: 180px; opacity: 0.5; }
+.v10-pca-owned-empty { display: flex; flex-direction: column; gap: 16px; }
+
+/* ----- Track-by-id disclosure ----- */
+.v10-pca-track { margin-top: 16px; }
+.v10-pca-track-toggle {
+  border: none;
+  background: transparent;
+  color: var(--text-link);
+  font-size: 12px;
+  cursor: pointer;
+  padding: 0;
+}
+.v10-pca-track-panel { display: flex; align-items: center; gap: 8px; margin-top: 8px; flex-wrap: wrap; }
+.v10-pca-track-panel .v10-form-input { max-width: 220px; }
+.v10-pca-track-status {
+  margin: 8px 0 0;
+  color: var(--text-secondary);
+  font-size: 12px;
+  line-height: 1.4;
+}
+
